requirements.txt
asgiref==3.2.10
Django==3.1.3
django-allauth==0.44.0
django-ckeditor==6.0.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-modeltranslation==0.16.1
django-ranged-response==0.2.0
django-recaptcha3==0.4.0
django-storages==1.11.1
django-simple-captcha==0.5.13
djangorestframework==3.12.2
pandas==1.2.4
Pillow==8.1.0
pipenv==2020.8.13
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.7.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1

